Register IngestGeoIpMetadata as a NamedXContent (#123079)
